+++ libnavigation-router/src/main/java/com/mapbox/navigation/route/internal/RouterWrapper.kt
@@ -188,6 +188,8 @@ class RouterWrapper(
             routeIndex,
             legIndex,
             RoutingProfile(routeOptions.profile().mapToRoutingMode(), routeOptions.user()),
+            routeOptions.baseUrl(),
+            accessToken,
             null
         )